feat: Introduce ability to add topics. Untested.
